+++ src/mongo_pipebuilder/__init__.py
@@ -9,6 +9,6 @@ Author: seligoroff
 
 from mongo_pipebuilder.builder import PipelineBuilder
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 __all__ = ["PipelineBuilder"]